Increase CursorWindow size to match AOSP; don't init lazily
Since we're not loading from a resource, we might as well initialize
eagerly; the compiler should be able to inline the constant instead
of having to do a bool check every time we a CursorWindow is created.

Loading the system resource would be nice
(com.android.internal.R.integer.config_cursorWindowSize) but it
doesn't seem that we have access to it.