getWindowSizeKb
The window size should ideally be an internal value. Alas, if a query
returns a column whose size is more than this window size, an
exception is raised. Thus, the only solution known currently is to
split the data in portions whose size are less than the actual window
size.

This can be done by looking at requery's code while splitting the
database content. However, I believe that being able to access this
value would be quite better, as it would allow the software using this
library to simply use `(int) (getWindowSizeKb * .9)` and ensure that
the data fits in the query, even if at some point this size change.

Ideally, I'd even prefer to be allowed to change this number when I
create a cursor, but I don't understand the codebase enough to know
how to do it safely, and how to test this library with the software I
contribute to (Ankidroid)